Forward Twitter secrets from host machine
The Docker host machine can set secrets locally and forward those values
to the same environment variables inside the application containers.
This way it is possible to omit the `config/twitter_*.json` files, that
are into the `.gitignore` and hence need additional setup after cloning
the repository.
This was already done for the database configuration; although without
forwarding local env-vars.
Using environment variables is more flexible and can be configured
independently from where the application is run and without the risk to
commit them on the (public) repository.
Using config files is still supported and they will be preferred, when
they are present on the filesystem.